requirements.txt
aiosignal==1.3.1
APScheduler==3.6.3
async-timeout==4.0.3
attrs==24.1.0
cachetools==4.2.2
certifi==2024.7.4
charset-normalizer==2.1.1
exceptiongroup==1.2.2
frozenlist==1.4.1
h11==0.14.0
idna==3.7
multidict==6.0.5
outcome==1.3.0.post0
PySocks==1.7.1
python-telegram-bot==13.12
pytz==2024.1
redis==5.0.8
requests==2.32.3
selenium==4.23.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
tornado==6.4.1
trio==0.26.1
trio-websocket==0.11.1
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.2
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.9.4
